Move filtering logic to Filter.js
